// src/gen/openapi.ts
export interface paths {
// ...
    "/action": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** @description Query for past executed actions. */
        get: {
            parameters: {
                query: {
                    /** @description Start action ID (exclusive) */
                    from: number;
                    /** @description End action ID (inclusive) */
                    to: number;
                    /** @description Optionally provided binary version of client, so nord server can ensure all actions provided are executable by client. */
                    clientVersion?: components["schemas"]["ExecutableVersion"] | null;
                };
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["ActionsItem"][];
                    };
                };
                400: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["RangeTooLarge"];
                    };
                };
                501: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["NotImplemented"];
                    };
                };
            };
        };
        put?: never;
        /** @description Send an action to nord to execute. This doesn't use `Protobuf<T>` since we want to length-delimit this message. `body` - (length prefix of action, action data protobuf, and signature). signature has exact length, but depend on action kind. body is exact payload used via other transports and journalled for rexecution */
        post: {
            parameters: {
                query?: never;
                header: {
                    "content-type": string;
                };
                path?: never;
                cookie?: never;
            };
            requestBody: {
                content: {
                    "application/octet-stream": unknown;
                };
            };
            responses: {
                /** @description byte stream */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/octet-stream": unknown;
                    };
                };
                413: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["PayloadTooLarge"];
                    };
                };
                415: {
                    headers: {
                        /**
                         * @description Expected request media type
                         * @example application/octet-stream
                         */
                        accept: string;
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["AcceptedMediaType"];
                    };
                };
            };
        };
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
// ...
        Trigger: {
            /** Format: uint32 */
            accountId: number;
            /** Format: uint32 */
            marketId: number;
            triggerPrice: components["schemas"]["PositivePriceMantissa"];
            limitPrice?: components["schemas"]["PositivePriceMantissa"] | null;
            side: components["schemas"]["Side"];
            kind: components["schemas"]["TriggerKind"];
            status: components["schemas"]["TriggerStatus"];
            createdAtActionId: string;
            finalizedAtActionId?: string | null;
            createdAt: string;
            finalizedAt: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}